requirements.txt
Django==1.11.15
djangorestframework==3.8.2
ipython==5.8.0
mysqlclient==1.3.12
django-push-notifications==1.6.0